showing the profile image in the nav
