+++ seeds/seed.js
@@ -1,6 +1,7 @@
 const sequelize = require('../config/connection');
 const { User, Project, Reaction, Social } = require('../models');
-
+const session = require('express-session');
+const SequelizeStore = require('connect-session-sequelize')(session.Store);
 const userData = require('./userData.json');
 const projectData = require('./projectData.json');
 const reactionData = require('./reactionData.json');
