+++ views/createproject.handlebars
@@ -1,0 +1,68 @@
+<div class='row mt-4'>
+	<div class='col-md-6'>
+		<h3>Create a New Project:</h3>
+
+		<form class='form new-project-form' enctype="multipart/form-data" >
+			<div class='form-group'>
+				<label for='project-name'>project name: *Needed</label>
+				<input
+					class='form-input'
+					type='text'
+					id='project-name'
+					name='project-name'
+				/>
+			</div>
+			<div class='form-group'>
+				<label for='project-github'>Github Link:</label>
+				<input
+					class='form-input'
+					id='project-github'
+					name='project-github'
+					placeholder="https://github.com/User/Project"
+				></input>
+			</div>
+			<div class='form-group'>
+				<label for='project-deployed'>Deployed Link:</label>
+				<input
+					class='form-input'
+					id='project-deployed'
+					name='project-deployed'
+					placeholder="https://DeployedLink.com"
+				></input>
+			</div>
+			<div class='form-group'>
+				<label for='project-caption'>caption: *Needed</label>
+				<input
+					class='form-input'
+					type='text'
+					id='project-caption'
+					name='project-caption'
+					placeholder='Something Captivating'
+				/>
+			</div>
+			<div class='form-group'>
+				<label for='project-desc'>description:</label>
+				<textarea
+					class='form-input'
+					id='project-desc'
+					name='project-desc'
+				></textarea>
+			</div>
+			
+			<div class="form-group">
+				<label for="project-image">Select Images:</label>
+				<input enctype="multipart/form-data" 
+				type="file" 
+				id="project-images" 
+				name="images"  
+				accept='image/png, image/gif, image/jpeg' multiple>
+			</div>
+			<div class="preview-container">
+				
+			</div>
+			<div class='form-group'>
+				<button type='submit' class='btn btn-primary'>Create</button>
+			</div>
+		</form>
+	</div>
+</div>
