+++ views/layouts/main.handlebars
@@ -27,7 +27,7 @@
 					Login
 				</button>
         {{else}}
-        <button onclick="window.location.href='/profile'">Profile</button>
+        <a class="profile_link" href="/profile"><img  class="profile-image-nav" src="{{profileImage}}" alt="" srcset=""></a>
         {{/unless}}
 			</nav>
 		</header>
